Add Notification Sound on Linux
This works out of the box with Ubuntu 16.04. This should work for most Linux distros with small modifications, but will only work out of the box with Ubuntu. The path of the notification sound is based on the default Ubuntu path to the Mallet.ogg file. (To-do: set path to ogg file in config.ini)